changes Proof of concept
The changes made as a proof as concept for the connection between the editor and the NLP model. Now the predictions of the model appear in the console log. To do this the user need first to select a chunk of text, the press on automatic annotation and then the annotations appear in the console log.